Trigger view key races actions on load
